feat(mega): implement http range request; general: impl multi-thread downloader to speed up some http request to remote storage. (e.g. Quark) general: process http Expires properly

+++ internal/model/args.go
@@ -28,7 +28,7 @@ type Link struct {
 	Expiration *time.Duration // local cache expire Duration
 	//for accelerating request, use multi-thread downloading
 	Concurrency int
-	PartSize    int64
+	PartSize    int
 }
 
 type OtherArgs struct {
